seperated request show view test from feature test
